requirements.txt
flask
pillow
imagehash
requests